Add files via upload
The LinkedAppliances.ino file contains the C code to to control two appliances (bulb and fan). The code will also raise an alarm and send an SMS when smoke is detected. Its a mini Smart Office Control System.
The login.htm is a sample login file. Please note that there is no real security with this file, it has solely been used for demonstration purposes
The panel.htm file has been used as a user interface to control the appliances
Good luck!